[Store] Align OffsetAllocatorBackendConfig with config layout
